// libs/admin-shared/src/model/forgetPwdEditRequest.ts
/**
 * Generated by orval v6.11.1 🍺
 * Do not edit manually.
 * 工业资源boss管理端【API接口文档】
 * 工业资源boss管理端
 */

/**
 * 忘记密码 ,编辑表单对象模型
 */
export interface ForgetPwdEditRequest {
  /** 确认新密码 */
  confirmPassword?: string;
  /** 登录账号 */
  loginName?: string;
  /** 新密码 */
  password?: string;
  /** 手机号 */
  phone?: string;
  /** 登录code,前端无需传,网关解析:接口生成已屏蔽 */
  tokenUserCode?: string;
  /** 登录tokenUserName,前端无需传,网关解析:接口生成已屏蔽 */
  tokenUserName?: string;
  /** 验证码 */
  verifyCode?: string;
}